Implement Trading UI For Meme Coin
- Refactor Meme Coin Page
- Add Trading UI
